format changes, deleted whatwedopage.tsx
